refactor: 게시글 추천 요청 record로 변경
