+++ src/main/java/net/detalk/api/controller/v1/request/CreateRecommend.java
@@ -1,0 +1,19 @@
+package net.detalk.api.controller.v1.request;
+
+import lombok.Builder;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+
+@Builder
+public record CreateRecommend (
+
+    @NotBlank(message = "Reason is required")
+    @Size(max = 255, message = "Reason must be less than or equal to 255 characters")
+    String reason,
+
+    @NotNull(message = "Member ID is required")
+    Long memberId
+
+) {}
