--- src/main/java/net/detalk/api/domain/CreateRecommend.java
@@ -1,17 +1,0 @@
-package net.detalk.api.domain;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class CreateRecommend {
-
-    private String reason;
-    private Long memberId;
-
-    @Builder
-    public CreateRecommend(String reason, Long memberId) {
-        this.reason = reason;
-        this.memberId = memberId;
-    }
-}
